refactor(dncl): split dncl-to-ruby and ruby-to-dncl into smaller modules
`dncl-to-ruby.js` (617 lines) と `ruby-to-dncl.js` (510 lines) が単一
ファイルとして大きくなりすぎていたため、責務単位で分割した。それぞれ
250 行を超えないように構成する。

dncl-to-ruby (DNCL→Ruby):

- dncl-keywords.js (74 lines) — DNCL_KEYWORDS, RUBY_LITERALS の定数
- dncl-state.js (118 lines) — arrayNames/functionNames の状態 +
  detectArrayNames / detectFunctionNames / mapVarName / accessor
- dncl-validator.js (49 lines) — validateDncl
- dncl-identifier-converter.js (193 lines) — convertIdentifier /
  convertIdentifiers / convertOperators / processSegments / 文字列変換
- dncl-builtins.js (53 lines) — convertBuiltinFunctions
- dncl-line-converter.js (173 lines) — convertLine + forLoopStack
- dncl-to-ruby.js (38 lines) — orchestration のみ

ruby-to-dncl (Ruby→DNCL):

- ruby-to-dncl-identifier.js (87 lines) — ID / convertVarRefs /
  convertOperators / processSegments
- ruby-to-dncl-builtins.js (134 lines) — replaceStandalonePCall /
  convertBuiltins
- ruby-to-dncl-line-converter.js (180 lines) — convertLine /
  detectForLoopPattern。blockStack はオーケストレータから受け取る
  パラメータに変更
- ruby-to-dncl.js (138 lines) — rubyToDncl orchestration

挙動の変化はなし。既存の DNCL 系 jest テスト 228 件すべて pass。
分割に伴い `forLoopStack` は state-resetter (`resetForLoopStack`) を
公開、`blockStack` は引数で渡す方式にして state を caller に集約した。

Refs #570 (review)
